Had to remove the SaveData class and move the functionality into the HomeViewController file. Data is now saving, and checking whether to see we have a saved user or not at launch
