+++ ProjectCaterpillar/Controllers/HomeViewController.swift
@@ -14,18 +14,17 @@ class HomeViewController: UIViewController, UIPickerViewDataSource, UIPickerView
     var user: User?
     var pet: Swallowtail?
     var lifeStages: [LifeStage] = []
-    let saveData = SaveData()
     
     @IBOutlet weak var userNameTextField: UITextField!
     
     @IBOutlet weak var petNameTextField: UITextField!
 
     @IBOutlet weak var homeScreenImage: UIImageView!
+        
     @IBAction func addButton(_ sender: UIButton) {
         collectUserData()
         guard let user = user else { return }
-        getDocumentsFolder()
-        saveData.save(information: user)
+        save(information: user)
         toggleDisplayData()
     }
     
@@ -60,8 +59,6 @@ class HomeViewController: UIViewController, UIPickerViewDataSource, UIPickerView
         pet?.name = petNameTextField.text!
         pet?.stageOfLife = lifeStages[lifeStagePicker.selectedRow(inComponent: 0)]
         user = User(name: newUserName, pet: pet!)
-//        print(user?.name)
-//        print(pet?.stageOfLife.name)
         
     }
     
@@ -85,41 +82,30 @@ class HomeViewController: UIViewController, UIPickerViewDataSource, UIPickerView
 
     }
     
-    func getDocumentsFolder() {
-        saveData.documentsDirectory()
-        saveData.dataFilePath()
-        print(saveData.dataFilePath())
-    }
-    
     func checkForLoadFile() -> Bool {
        let fileManager = FileManager()
-       let filePath = saveData.dataFilePath().path
+       let filePath = dataFilePath().path
        
        return fileManager.fileExists(atPath: filePath)
-    
     }
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        let checkForFile = checkForLoadFile()
+        populateLifeStage()
         
-        print(checkForLoadFile())
-
-        if checkForLoadFile() {
-            guard let user = user else { return }
-            saveData.load(for: user)
-            print(user)
+        if checkForFile {
+            //MARK: TODO: From saveData.load, user is nil
+            load()
+            print(pet)
             toggleDisplayData()
         } else {
             loadImage()
             nameDisplayLabel.isHidden = true
             petNameDisplayLabel.isHidden = true
+            var swallowTail = Swallowtail(name: "", startDate: "", imgFileName: "", stageOfLife: lifeStagesDatabase.lifestages[0])
+            pet = swallowTail
         }
-        
-        populateLifeStage()
-        
-        var swallowTail = Swallowtail(name: "", startDate: "", imgFileName: "", stageOfLife: lifeStagesDatabase.lifestages[0])
-        pet = swallowTail
-        
     }
 
     override func didReceiveMemoryWarning() {
@@ -127,4 +113,41 @@ class HomeViewController: UIViewController, UIPickerViewDataSource, UIPickerView
 
     }
 
+}
+
+// MARK: Save and Load user functionality
+
+extension HomeViewController {
+    
+    func documentsDirectory() -> URL {
+        let paths = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)
+        return paths[0]
+    }
+    
+    func dataFilePath() -> URL {
+        return documentsDirectory().appendingPathComponent("SavedData.plist")
+    }
+    
+    func save(information: User) {
+        let encoder = PropertyListEncoder()
+        do {
+            let data = try encoder.encode(information)
+            try data.write(to: dataFilePath(), options: Data.WritingOptions.atomic)
+        } catch {
+            print("Error saving data")
+        }
+    }
+    
+    func load() {
+        let path = dataFilePath()
+        if let data = try? Data(contentsOf: path) {
+            let decoder = PropertyListDecoder()
+            do {
+                user = try decoder.decode(User.self, from: data)
+                pet = user?.pet
+            } catch {
+                print("Error decoding object")
+            }
+        }
+    }
 }
